修改滑动标题栏

+++ iosapp/EmojiPanelVC.m
@@ -69,7 +69,7 @@
     UICollectionViewCell *cell = [collectionView dequeueReusableCellWithReuseIdentifier:@"EmojiCell" forIndexPath:indexPath];
     
     NSInteger emojiNum = _pageIndex * 21 + indexPath.section * 7 + indexPath.row + 1;
-    NSString *emojiImageName = [NSString stringWithFormat:@"%03ld", emojiNum];
+    NSString *emojiImageName = [NSString stringWithFormat:@"%03ld", (long)emojiNum];
     [cell setBackgroundView:[[UIImageView alloc] initWithImage:[UIImage imageNamed:emojiImageName]]];
     
     return cell;
@@ -78,7 +78,7 @@
 - (void)collectionView:(UICollectionView *)collectionView didSelectItemAtIndexPath:(NSIndexPath *)indexPath
 {
     NSInteger emojiNum = _pageIndex * 21 + indexPath.section * 7 + indexPath.row + 1;
-    NSString *emojiImageName = [NSString stringWithFormat:@"%03ld", emojiNum];
+    NSString *emojiImageName = [NSString stringWithFormat:@"%03ld", (long)emojiNum];
     
     NSTextAttachment *textAttachment = [NSTextAttachment new];
     textAttachment.image = [UIImage imageNamed:emojiImageName];
